Update spec to account for rpkg in copr
With the new way of publishing in copr, we
are using rpkg which seems to build Source0 from the current dir.

This is an issue, as npm packages (the current Source0)
are published with the top-level dir having the name
package.

This causes a variant between how rpkg packages source,
and npm's packaged source.

To work around this for now, force Source0 to be local.

Signed-off-by: Joe Grund <joe.grund@intel.com>