Use Thot to encode / decode types
It looks like Fable 2 will drop automagic serialization in favor of
explicit serialization / deserialization.

It also appears that it will introduce generation
of encoding / decoding code at some point in the future.

In the interim, we should move to creating our own encoding / decoding members
and add sufficient tests to ensure they are working as expected.

  - Add .editorconfig.
  - Add test to travis.
  - Reorg Fable.Import.NodeLibzfs for source
     and test.
  - Bump F# deps.
  - Add jest testing.
  - Bump version to 0.2.0

Signed-off-by: Joe Grund <joe.grund@intel.com>