Move bindings into src
We'd like to have completion via RLS + documentation
published to crates.io.

This is not possible with dynamic includes with the current out
directory via bindgen.

Instead, commit the bindings into SCM. Longer term,
we can look at generating bindings for multiple versions of ZFS / OSes
but we don't need to tackle this now.

Signed-off-by: Joe Grund <joe.grund@intel.com>