Enumerate zfs props
Fixes #44.

In order to use a single shared structure from Fable.Import.NodeLibzfs, we should add the ability enumerate all properties from zfs objects and zpools.

For now, we can just enumerate the zfs objects as that's what we are currently interested in, and just push an empty array for the zpool object properties.

Signed-off-by: Joe Grund <joe.grund@intel.com>