Add guid prop to dataset + encode / decode + getDs fn
Add the guid to the top level dataset objects returned.

We need the ds guid, because that's how IML stores it in
corosync.

In addition, add a fn to fetch dataset by name so we
can get the relevant data without pulling all imported pools.

Finally, add thot encode / decode to pools, datasets and props
so we can have a singular point to encode / decode.

Signed-off-by: Joe Grund <joe.grund@intel.com>