Add fn to encode to fable from JS.
Add a method to encode the JS VDev
tree received from node-libzfs
into a Fable recursive DU.

This needs to be done because fable cannot
understand an incoming JS object as a DU
when it's missing the tags fable needs.

This function will pass over the VDev tree
once to convert it.

Signed-off-by: Joe Grund <joe.grund@intel.com>